move node-installer testdata into sub directory for clarity
