+++ testdata/node-installer/assets/containerd-shim-slight-v1
