+++ testdata/node-installer/assets/containerd-shim-spin-v1
