+++ testdata/node-installer/containerd/existing-containerd-shim-config/etc/containerd/config.toml
