+++ testdata/node-installer/containerd/existing-containerd-shim-config/opt/kwasm/bin/containerd-shim-spin-v1
