+++ testdata/node-installer/containerd/existing-containerd-shim-config/opt/kwasm/kwasm-lock.json
