+++ testdata/node-installer/containerd/missing-containerd-shim-config/etc/containerd/config.toml
