+++ testdata/node-installer/shim/opt/kwasm/bin/containerd-shim-spin-v1
