+++ testdata/node-installer/shim/opt/kwasm/kwasm-lock.json
