+++ tests/node-installer/fs.go
